+++ Models/Task/Task.h
@@ -24,10 +24,10 @@ private:
     float ram;
 
     //Megabytes
-    unsigned int data_in;
+    float data_in;
 
     //Megabytes
-    unsigned int data_out;
+    float data_out;
 
     //Megabytes
     float storage;
@@ -43,8 +43,8 @@ private:
     float process_time_mobile;
 
 public:
-    Task(const std::string &name, float process_time_cloud, float process_time_edge, float process_time_mobile, float ram, unsigned int dataIn,
-         unsigned int dataOut, float storage, int source_mobile_id, bool can_offload);
+    Task(const std::string &name, float process_time_cloud, float process_time_edge, float process_time_mobile, float ram, float dataIn,
+         float dataOut, float storage, int source_mobile_id, bool can_offload);
 
     int getId() const;
 
@@ -68,13 +68,13 @@ public:
 
     void setRam(float ram);
 
-    unsigned int getDataIn() const;
+    float getDataIn() const;
 
-    void setDataIn(unsigned int dataIn);
+    void setDataIn(float dataIn);
 
-    unsigned int getDataOut() const;
+    float getDataOut() const;
 
-    void setDataOut(unsigned int dataOut);
+    void setDataOut(float dataOut);
 
     float getStorage() const;
